feat: ios: unify Rust services uasge of BackendService
