+++ ios/PolkadotVault/RustDataExtensions/Address+DisplayablePath.swift
