+++ ios/PolkadotVault/RustDataExtensions/Encryption+RawRepresentable.swift
