+++ ios/PolkadotVault/RustDataExtensions/MKeyDetails+Helpers.swift
