+++ ios/PolkadotVault/RustDataExtensions/QrData+Helpers.swift
