+++ ios/PolkadotVault/RustDataExtensions/Verifier+Show.swift
