+++ ios/PolkadotVault/RustDataExtensions/VerifierValue+Show.swift
